# ENTREGABLES_SIGAI_SES/01_DOCUMENTACION_TECNICA/API_SPEC/openapi.yaml
openapi: 3.0.3
info:
  title: SIGAI-SES API
  version: "1.0.0"
  description: >
    API RESTful del Sistema Integral de Gestión de Activos e Inventario
    para Seguridad Electrónica Securitas Colombia.
  contact:
    name: Equipo de Desarrollo SIGAI-SES
    email: soporte@ses-securitas.com.co
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.ses.example.com/v1
    description: Producción
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido de /auth/login
  schemas:
...
    Item:
      type: object
      properties:
        id_item:
          type: integer
        categoria:
          type: string
          enum: [MONITOREO, MANTENIMIENTO, INSTALACION, SOLUCIONES, EPP, CONSUMIBLE, HERRAMIENTA_LAB, REPUESTO]
        nombre_equipo:
          type: string
        marca:
          type: string
          nullable: true
        referencia:
          type: string
        codigo_item_interno:
          type: string
        stock_minimo:
          type: integer
        costo_unitario:
          type: number
          format: decimal
        moneda:
          type: string
          enum: [COP, USD, EUR]
        created_at:
          type: string
          format: date-time
    ItemCreate:
      type: object
      required: [categoria, nombre_equipo]
      properties:
        categoria:
          type: string
        nombre_equipo:
          type: string
        marca:
          type: string
        referencia:
          type: string
        codigo_item_interno:
          type: string
        stock_minimo:
          type: integer
          default: 5
        compra_maxima:
          type: integer
          default: 20
        costo_unitario:
          type: number
          default: 0.00
        moneda:
          type: string
          default: COP
        cantidad_inicial:
          type: number
          description: Stock inicial para crear el registro bulk

    # ==================== ACTIVOS ====================
    Activo:
      type: object
      properties:
        id_activo:
          type: integer
        id_item:
          type: integer
        serial:
          type: string
        estado_actual:
          type: string
          enum: [DISPONIBLE, INSTALADO, EN_GARANTIA, REPARADO, LABORATORIO, DESMONTE, BAJA, OBSOLETO]
        condicion_fisica:
          type: string
          enum: [NUEVO, USADO_BUENO, PARA_REPARAR, SULFATADO, SIN_CONTRAPESOS, DANADO]
        ubicacion_fisica:
          type: string
        area_asignada:
          type: string
        responsable_sitio:
          type: string
        activo_fijo_securitas:
          type: string
        calificacion_tecnica:
          type: string
          enum: [BUENO, RECUPERABLE, DESECHO]
        updated_at:
          type: string
          format: date-time
    ActivoCreate:
      type: object
      required: [id_item, serial]
      properties:
        id_item:
          type: integer
        serial:
          type: string
        estado_actual:
          type: string
          default: DISPONIBLE
        condicion_fisica:
          type: string
          default: NUEVO
        ubicacion_fisica:
          type: string
        area_asignada:
          type: string
        responsable_sitio:
          type: string
        id_proyecto_actual:
          type: integer
        id_proveedor_compra:
          type: integer
        numero_factura_compra:
          type: string
        fecha_compra:
          type: string
          format: date
        activo_fijo_securitas:
          type: string
        observaciones:
          type: string
    DesmonteBulkCreate:
      type: object
      required: [id_item, seriales]
      properties:
        id_item:
          type: integer
        id_proyecto:
          type: integer
        id_cliente:
          type: integer
        seriales:
          type: array
          items:
            type: string

    # ==================== CLIENTES ====================
...
security:
  - BearerAuth: []
paths:
...
  /inventory/items:
    get:
      summary: Listar items del catalogo
      tags: [Inventory]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Lista de items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      summary: Crear nuevo item
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemCreate"
      responses:
        "201":
          description: Item creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /inventory/items/{id_item}:
    get:
      summary: Obtener item por ID
      tags: [Inventory]
      parameters:
        - name: id_item
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Datos del item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    put:
      summary: Actualizar item
      tags: [Inventory]
      parameters:
        - name: id_item
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemCreate"
      responses:
        "200":
          description: Item actualizado
    delete:
      summary: Eliminar item (soft delete)
      tags: [Inventory]
      parameters:
        - name: id_item
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item eliminado
  /inventory/items/desmonte-bulk:
    post:
      summary: Registrar desmontes masivos
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesmonteBulkCreate"
      responses:
        "201":
          description: Desmontes registrados
  /inventory/activos:
    get:
      summary: Listar activos
      tags: [Inventory]
      parameters:
        - name: estado
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Lista de activos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activo"
    post:
      summary: Crear activo
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivoCreate"
      responses:
        "201":
          description: Activo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activo"
  /inventory/activos/{serial}:
    get:
      summary: Obtener activo por serial
      tags: [Inventory]
      parameters:
        - name: serial
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Datos del activo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activo"
  /inventory/activos/{id_activo}:
    put:
      summary: Actualizar activo
      tags: [Inventory]
      parameters:
        - name: id_activo
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activo actualizado
    delete:
      summary: Eliminar activo
      tags: [Inventory]
      parameters:
        - name: id_activo
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Activo eliminado
  /inventory/activos/{id_activo}/triaje:
    patch:
      summary: Actualizar calificacion tecnica (triaje)
      tags: [Inventory]
      parameters:
        - name: id_activo
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                calificacion_tecnica:
                  type: string
                  enum: [BUENO, RECUPERABLE, DESECHO]
                observaciones:
                  type: string
      responses:
        "200":
          description: Triaje actualizado
  /inventory/activos/{id_activo}/historial:
    get:
      summary: Historial de ubicaciones del activo
      tags: [Inventory]
      parameters:
        - name: id_activo
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Historial de ubicaciones
...